Revised code and started just spawning one thread at a time to confirm I am on the right track
Code seems to be correctly spawning new machine threads, and the machine thread is correctly appending itself to the machine process list, which is started and joined.

The printQueue seems to be working fine, but need to download the original code now that the assignment briefing has been changed to take from the head of the queue as opposed to the tail.

Machine ID needs to be fixed in console output as it current reads "Machine <built=in function id> Sent a print request"

Seems to be correctly counting the number of processes in the queue, and capping at the maximum number of printers available